app-local/sync-agent/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
playwright==1.40.0
